+++ retail/interfaces/clients/aws_lambda/client.py
@@ -1,4 +1,12 @@
-from typing import Protocol, Dict, Any
+from dataclasses import dataclass
+from typing import Any, Dict, Protocol
+
+
+@dataclass
+class RequestData:
+    params: dict
+    payload: dict
+    credentials: dict
 
 
 class AwsLambdaClientInterface(Protocol):
@@ -8,4 +16,7 @@ class AwsLambdaClientInterface(Protocol):
     def update_function_code(
         self, function_name: str, zip_bytes: bytes
     ) -> Dict[str, Any]:
+        ...
+
+    def invoke(self, function_name: str, data: RequestData) -> Dict[str, Any]:
         ...